% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Az.r
\name{Az}
\alias{Az}
\title{estimate minimum cutting diameter}
\usage{
Az(sp, dbh)
}
\arguments{
\item{sp}{Bdat species code [1;36], integer}

\item{dbh}{vector of diameter in breast height, numeric}
}
\value{
vector of minimum cutting diameter [cm].
}
\description{
estimate minimum cutting diameter based on diameter in breast
height based on the functions implemented in BDAT
}
\details{
the implemented BDAT function and parameters are used. Not all
BDAT-species possess their own parameters, hence most of them are matched to
one of the main tree species, especially in deciduous tree species (only
parameters for beech and oak are available).
}
\examples{
sp <- 1
dbh <- 30
Az(sp, dbh)
}
