% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{biomass}
\alias{biomass}
\title{Prediction of above-ground biomass according to NFI-functions}
\usage{
biomass(spp, d13, d03, h)
}
\arguments{
\item{spp}{vector of species code for biomass function [1;18]}

\item{d13}{vector of diameter in breast height in centimeter}

\item{d03}{vector of diameter in 30\% of tree height in centimeter}

\item{h}{vector of height of trees}
}
\value{
a vector of total above-ground biomass
}
\description{
Prediction of total above-ground biomass for trees defined via species, dbh,
d03 and height
}
\details{
code taken from BDAT (Koeff.f).
}
\references{
Riedel, T. and G. Kaendler (2017). "Nationale Treibhausgasberichterstattung:
Neue Funktionen zur Schätzung der oberirdischen Biomasse am Einzelbaum."
Forstarchiv 88(2): 31-38.
}
