% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEandVCOV.r
\name{calcVCOVsekVol}
\alias{calcVCOVsekVol}
\title{calculate VCOV-Matrix for volume segments}
\usage{
calcVCOVsekVol(estD, kovD, estL)
}
\arguments{
\item{estD}{vector of estimated diameter, numeric}

\item{kovD}{variance-covariance-matrix of the estimated diameter, numeric}

\item{estL}{vector of segment length, numeric}
}
\value{
variance-covariance matrix of the segment volume
}
\description{
calculate variance-covariance matrix for volume segments from
the estimated diameter and uncertainty information from TapeR-model
}
\details{
Calculations according to rules for products and sums of variances
}
