% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnUnvd.r
\name{fnUnvd}
\alias{fnUnvd}
\title{percentage of unusable coarse wood}
\usage{
fnUnvd(ba = NULL, dm = NULL, cd = NULL)
}
\arguments{
\item{ba}{tree species index; see details}

\item{dm}{diameter class; see details}

\item{cd}{cutting diameter; see details}
}
\description{
function extracts the percentage of unusable coarse wood
according to species (beech, oak), diameter class and cutting diameter
}
\details{
Function extracts the percentage of unusable coarse wood according
to three parameters: (i) tree species, which is 1 for using beech models and
2 for using the oak model; (ii) the 2cm-diameter class (from 8 and 60cm) and
(iii) the cutting diameter ranging from 8 to 40cm.
}
\references{
Kublin and Scharnagl (1988): Verfahrens- und Programmbeschreibung
zum BWI-Unterprogramm BDAT. FVA-BW 1988. ISSN: 0178-3165.
}
