% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estHeight.r
\name{HtCoef}
\alias{HtCoef}
\title{returns coefficients for Pettersen-Height model}
\usage{
HtCoef(sp = NULL, qtl = NULL)
}
\arguments{
\item{sp}{BDAT species code, could be NULL then all coefficients are returned}

\item{qtl}{quantile, either NULL or 17, 50, 83}
}
\value{
a data.frame with species code and coefficients
}
\description{
Function to provide model coefficients for Pettersen-height model
}
