% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmeSKEBLUP}
\alias{lmeSKEBLUP}
\title{diameter prediction E[d] for TapeR-object}
\usage{
lmeSKEBLUP(xm, ym, xp, par, RV)
}
\arguments{
\item{xm}{relative height of measured diameter}

\item{ym}{measured diameter for calibration}

\item{xp}{relative height for which diameter prediction is required}

\item{par}{a TapeR-object (including padded knots vector)}

\item{RV}{numeric vector holding assumed residual variance for each observation}
}
\value{
a list holding several elements, perspectively only the estimated diameter
}
\description{
Prediction diameter (no variances) for given tree and TapeR-object using
BSpline Matrix all in C++
}
\details{
code implementation in C++ following the code base of TapeR. Bspline
  matrix code taken from R-package splines to avoid the need of calling R from C.
}
