% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{petterson}
\alias{petterson}
\title{height estimation}
\usage{
petterson(sp, d13)
}
\arguments{
\item{sp}{vector of species code for biomass function from interval [1;18];
see \code{\link{BaMap}} for mapping of species model codes}

\item{d13}{vector of diameter in breast height; in centimeter}
}
\value{
a scalar: tree height
}
\description{
height estimation based on diameter in breast height and species using a
Petterson-function
}
