% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiPar.r
\name{RiPar}
\alias{RiPar}
\title{parameter of bark function}
\usage{
RiPar(ba = NULL, fn = NULL, par = NULL)
}
\arguments{
\item{ba}{tree species code; returned by \code{\link{BaMap}}}

\item{fn}{function number; see details}

\item{par}{parameter; see details}
}
\description{
extract parameter of bark functions according to Altherr et al.
1974 - 1979
}
\details{
Function extracts the parameter according to tree species, function
type and parameter number. There are three parameters in each of four
functions. The first one refers to butt log (dt. Erdstamm), the second to
middle log (dt. Mittelstammstück), the third to the top log
(dt. Gipfelstammstück) and the fourth to the complete stem (dt. Gesamtstamm).
}
