% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreesAccessors.r
\name{Slot_accessors}
\alias{Slot_accessors}
\alias{spp}
\alias{spp,tprTrees-method}
\alias{spp<-}
\alias{spp<-,tprTrees-method}
\alias{Dm}
\alias{Dm,tprTrees-method}
\alias{Dm<-}
\alias{Dm<-,tprTrees-method}
\alias{Hm}
\alias{Hm,tprTrees-method}
\alias{Hm<-}
\alias{Hm<-,tprTrees-method}
\alias{Ht}
\alias{Ht,tprTrees-method}
\alias{Ht<-}
\alias{Ht<-,tprTrees-method}
\alias{sHt}
\alias{sHt,tprTrees-method}
\alias{sHt<-}
\alias{sHt<-,tprTrees-method}
\alias{mono}
\alias{mono,tprTrees-method}
\title{slot accessor functions for class 'tprtrees'}
\usage{
spp(obj)

\S4method{spp}{tprTrees}(obj)

spp(obj) <- value

\S4method{spp}{tprTrees}(obj) <- value

Dm(obj)

\S4method{Dm}{tprTrees}(obj)

Dm(obj) <- value

\S4method{Dm}{tprTrees}(obj) <- value

Hm(obj)

\S4method{Hm}{tprTrees}(obj)

Hm(obj) <- value

\S4method{Hm}{tprTrees}(obj) <- value

Ht(obj)

\S4method{Ht}{tprTrees}(obj)

Ht(obj) <- value

\S4method{Ht}{tprTrees}(obj) <- value

sHt(obj)

\S4method{sHt}{tprTrees}(obj)

sHt(obj) <- value

\S4method{sHt}{tprTrees}(obj) <- value

mono(obj)

\S4method{mono}{tprTrees}(obj)
}
\arguments{
\item{obj}{object of class 'tprtrees'}

\item{value}{depending on slot, see details}
}
\value{
the accessor functions return the value of the specified slot and
the setting functions update the object
}
\description{
get and set slot values
}
\details{
Getting and setting the values of the different slots of
'tprTrees'-objects. For slot \code{mono} no setting function has been defined,
as this slot is computed by \code{\link{check_monotonicity}} and should not
be reset by users.

Setting of \code{spp} requires mode integer. For convenience,
\code{value} is coerced by \code{\link{as.integer}}.

Setting \code{spp} and \code{H}, a vector of length equal
\code{length(spp(obj))} is required.

For setting slots \code{Dm} and \code{Hm} \code{value} must be a
list of vectors of length equal \code{length(spp(obj))} and the length of
each vector must correspond to the length of the vectors in \code{Hm} and
\code{Dm}.
}
\section{Functions}{
\itemize{
\item \code{spp()}: getting slot 'spp' of \code{obj}

\item \code{spp(tprTrees)}: method for class 'tprTrees'

\item \code{spp(obj) <- value}: setting 'spp' slot of object

\item \code{spp(tprTrees) <- value}: method for class 'tprTrees'

\item \code{Dm()}: getting slot 'Dm' of \code{obj}

\item \code{Dm(tprTrees)}: method for class 'tprTrees'

\item \code{Dm(obj) <- value}: setting 'Dm' slot of object

\item \code{Dm(tprTrees) <- value}: method for class 'tprTrees'

\item \code{Hm()}: getting slot 'Hm' of \code{obj}

\item \code{Hm(tprTrees)}: method for class 'tprTrees'

\item \code{Hm(obj) <- value}: setting 'Hm' slot of object

\item \code{Hm(tprTrees) <- value}: method for class 'tprTrees'

\item \code{Ht()}: getting slot 'Ht' of \code{obj}

\item \code{Ht(tprTrees)}: method for class 'tprTrees'

\item \code{Ht(obj) <- value}: setting 'Ht' slot of object

\item \code{Ht(tprTrees) <- value}: method for class 'tprTrees'

\item \code{sHt()}: getting slot 'sHt' of \code{obj}

\item \code{sHt(tprTrees)}: method for class 'tprTrees'

\item \code{sHt(obj) <- value}: setting 'sHt' slot of object

\item \code{sHt(tprTrees) <- value}: method for class 'tprTrees'

\item \code{mono()}: getting slot 'monotone' of \code{obj}

\item \code{mono(tprTrees)}: method for class 'tprTrees'

}}
\seealso{
\code{\link{tprTrees-class}}, \code{\link{tprTrees}}
}
