% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TapeS-package.r
\docType{package}
\name{TapeS-package}
\alias{-package}
\alias{TapeS-package}
\title{TapeS: New taper functions for most common tree species in Germany and
  evaluating functions.}
\description{
Implementing new taper functions for German NFI and making available
associated function to evaluate diameter, height of diameter, bark thickness,
biomass and volume of these taper functions. Also providing wrappers to
calculate specific diameters like dbh, d005, d03 and d7 as well as specific
volume measures like Vfm, Efm, VolR, VolE, Vol_FAO, and physical stem volume.
}
\details{
The package contains datasets which hold the taper curve models and further
parameters used inside the provided functions
(see \code{data(package="TapeS")}).
}
\keyword{package}
