% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TapeS-package.r
\docType{package}
\name{TapeS-package}
\alias{TapeS}
\alias{TapeS-package}
\title{TapeS: Tree Taper Curves and Sorting Based on 'TapeR'}
\description{
Providing new german-wide 'TapeR' Models and functions for their evaluation. Included are the most common tree species in Germany (Norway spruce, Scots pine, European larch, Douglas fir, Silver fir as well as European beech, Common/Sessile oak and Red oak). Many other species are mapped to them so that 36 tree species / groups can be processed. Single trees are defined by species code, one or multiple diameters in arbitrary measuring height and tree height. The functions then provide information on diameters along the stem, bark thickness, height of diameters, volume of the total or parts of the trunk and total and component above-ground biomass. It is also possible to calculate assortments from the taper curves. For diameter and volume estimation, uncertainty information is given.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/vochr/tapes}
}

}
\author{
\strong{Maintainer}: Christian Vonderach \email{christian.vonderach@forst.bwl.de}

Authors:
\itemize{
  \item Edgar Kublin
  \item Gerald Kändler
}

}
\keyword{internal}
